move land entry to entity
