+++ entity/land_entry.go
@@ -1,12 +1,10 @@
-package landprovider
-
-import "github.com/uber/submitqueue/entity"
+package entity
 
 // LandEntry pairs a land strategy with the change to land.
 // Each entry represents one request's contribution to a batch land operation.
 type LandEntry struct {
 	// Strategy is the source control integration method for this change.
-	Strategy entity.RequestLandStrategy
+	Strategy RequestLandStrategy
 	// Change is the code change to land.
-	Change entity.Change
+	Change Change
 }
